backend/requirements.txt
Flask
flask-cors
geopy
requests
python-dotenv
